Renamed hpref_serial/hpref to hpref/phpref in miniapps/meshing
