+++ miniapps/meshing/CMakeLists.txt
@@ -93,8 +93,8 @@ add_mfem_miniapp(polar-nc
   MAIN polar-nc.cpp
   LIBRARIES mfem)
 
-add_mfem_miniapp(hpref_serial
-  MAIN hpref_serial.cpp
+add_mfem_miniapp(hpref
+  MAIN hpref.cpp
   LIBRARIES mfem)
 
 # Add serial tests.
@@ -112,7 +112,7 @@ if (MFEM_ENABLE_TESTING)
     reflector
     ref321
     mesh-quality
-    hpref_serial
+    hpref
   )
   foreach(test ${SERIAL_TESTS})
     add_test(NAME ${test} COMMAND ${test} -no-vis)
@@ -143,8 +143,8 @@ if (MFEM_USE_MPI)
     MAIN pminimal-surface.cpp
     LIBRARIES mfem)
 
-  add_mfem_miniapp(hpref
-    MAIN hpref.cpp
+  add_mfem_miniapp(phpref
+    MAIN phpref.cpp
     LIBRARIES mfem)
 
   # Add parallel tests.
@@ -154,7 +154,7 @@ if (MFEM_USE_MPI)
       pmesh-fitting
       fit-node-position
       pminimal-surface
-      hpref
+      phpref
     )
     # Meshing miniapps that return MFEM_SKIP_RETURN_VALUE in some cases:
     set(SKIP_TESTS)
