+++ miniapps/meshing/hpref.cpp
@@ -1,19 +1,19 @@
-//                       Parallel hp-refinement example
+//                       Serial hp-refinement example
 //
 // Compile with: make hpref
 //
-// Sample runs:  mpirun -np 4 hpref -dim 2 -n 1000
-//               mpirun -np 8 hpref -dim 3 -n 200
+// Sample runs:  hpref -dim 2 -n 1000
+//               hpref -dim 3 -n 500
 //
-// Description:  This example demonstrates h- and p-refinement in a parallel
-//               finite element discretization of the Poisson problem (cf. ex1p)
+// Description:  This example demonstrates h- and p-refinement in a serial
+//               finite element discretization of the Poisson problem (cf. ex1)
 //               -Delta u = 1 with homogeneous Dirichlet boundary conditions.
 //               Refinements are performed iteratively, each iteration having h-
-//               or p-refinements on all MPI processes. For simplicity, we
-//               randomly choose the elements and the type of refinement, for
-//               each iteration. In practice, these choices may be made in a
-//               problem-dependent way, but this example serves only to
-//               illustrate the capabilities of hp-refinement in parallel.
+//               or p-refinements. For simplicity, we randomly choose the
+//               elements and the type of refinement, for each iteration. In
+//               practice, these choices may be made in a problem-dependent way,
+//               but this example serves only to illustrate the capabilities of
+//               hp-refinement.
 
 //               We recommend viewing Example 1 before viewing this example.
 
@@ -24,7 +24,7 @@
 using namespace std;
 using namespace mfem;
 
-real_t CheckH1Continuity(ParGridFunction & x);
+real_t CheckH1Continuity(GridFunction & x);
 
 // Deterministic function for "random" integers.
 int DetRand(int & seed)
@@ -37,13 +37,7 @@ void f_exact(const Vector &x, Vector &f);
 
 int main(int argc, char *argv[])
 {
-   // 1. Initialize MPI and HYPRE.
-   Mpi::Init();
-   const int num_procs = Mpi::WorldSize();
-   const int myid = Mpi::WorldRank();
-   Hypre::Init();
-
-   // 2. Parse command-line options.
+   // 1. Parse command-line options.
    int order = 1;
    const char *device_config = "cpu";
    bool visualization = true;
@@ -69,27 +63,20 @@ int main(int argc, char *argv[])
    args.AddOption(&projectSolution, "-proj", "--project-solution", "-no-proj",
                   "--no-project",
                   "Whether to project a coefficient to solution");
-
    args.Parse();
    if (!args.Good())
    {
-      if (myid == 0)
-      {
-         args.PrintUsage(cout);
-      }
+      args.PrintUsage(cout);
       return 1;
    }
-   if (myid == 0)
-   {
-      args.PrintOptions(cout);
-   }
+   args.PrintOptions(cout);
 
-   // 3. Enable hardware devices such as GPUs, and programming models such as
+   // 2. Enable hardware devices such as GPUs, and programming models such as
    //    CUDA, OCCA, RAJA and OpenMP based on command line options.
    Device device(device_config);
-   if (myid == 0) { device.Print(); }
+   device.Print();
 
-   // 4. Construct a uniform coarse mesh on all processors.
+   // 3. Construct a uniform coarse mesh on all processors.
    Mesh mesh;
    if (dim == 3)
    {
@@ -102,22 +89,9 @@ int main(int argc, char *argv[])
 
    mesh.EnsureNCMesh();
 
-   // 5. Define a parallel mesh by a partitioning of the serial mesh. Refine
-   //    this mesh further in parallel to increase the resolution. Once the
-   //    parallel mesh is defined, the serial mesh can be deleted.
-   ParMesh pmesh(MPI_COMM_WORLD, mesh);
-   mesh.Clear();
-   {
-      int par_ref_levels = 0;
-      for (int l = 0; l < par_ref_levels; l++)
-      {
-         pmesh.UniformRefinement();
-      }
-   }
-
-   // 6. Define a parallel finite element space on the parallel mesh. Here we
-   //    use continuous Lagrange finite elements of the specified order. If
-   //    order < 1, we instead use an isoparametric/isogeometric space.
+   // 4. Define a finite element space on the mesh. Here we use continuous
+   //    Lagrange finite elements of the specified order. If order < 1, we
+   //    instead use an isoparametric/isogeometric space.
    FiniteElementCollection *fec;
    bool delete_fec;
    if (order > 0)
@@ -125,14 +99,11 @@ int main(int argc, char *argv[])
       fec = new H1_FECollection(order, dim);
       delete_fec = true;
    }
-   else if (pmesh.GetNodes())
+   else if (mesh.GetNodes())
    {
-      fec = pmesh.GetNodes()->OwnFEC();
+      fec = mesh.GetNodes()->OwnFEC();
       delete_fec = false;
-      if (myid == 0)
-      {
-         cout << "Using isoparametric FEs: " << fec->Name() << endl;
-      }
+      cout << "Using isoparametric FEs: " << fec->Name() << endl;
    }
    else
    {
@@ -141,13 +112,12 @@ int main(int argc, char *argv[])
    }
 
    const int fespaceDim = projectSolution ? dim : 1;
-   ParFiniteElementSpace fespace(&pmesh, fec, fespaceDim);
+   FiniteElementSpace fespace(&mesh, fec, fespaceDim);
 
-   // 7. Iteratively perform h- and p-refinements.
-
+   // 5. Iteratively perform h- and p-refinements.
    int numH = 0;
    int numP = 0;
-   int seed = myid;
+   int seed = 0;
 
    const std::vector<char> hp_char = {'h', 'p'};
 
@@ -155,13 +125,11 @@ int main(int argc, char *argv[])
    {
       const int r1 = deterministic ? DetRand(seed) : rand();
       const int r2 = deterministic ? DetRand(seed) : rand();
-      const int elem = r1 % pmesh.GetNE();
-      int hp = r2 % 2;
-      MPI_Bcast(&hp, 1, MPI_INT, 0, MPI_COMM_WORLD);
+      const int elem = r1 % mesh.GetNE();
+      const int hp = r2 % 2;
 
-      if (myid == 0)
-         cout << "hp-refinement iteration " << iter << ": "
-              << hp_char[hp] << "-refinement" << endl;
+      cout << "hp-refinement iteration " << iter << ": "
+           << hp_char[hp] << "-refinement" << endl;
 
       if (hp == 1)
       {
@@ -176,23 +144,21 @@ int main(int argc, char *argv[])
          // h-ref
          Array<Refinement> refs;
          refs.Append(Refinement(elem));
-         pmesh.GeneralRefinement(refs);
+         mesh.GeneralRefinement(refs);
          fespace.Update(false);
          numH++;
       }
    }
 
-   const HYPRE_BigInt size = fespace.GlobalTrueVSize();
-   const int maxP = fespace.GetMaxElementOrder();
-   if (myid == 0)
-   {
-      cout << "Number of finite element unknowns: " << size << endl;
-      cout << "Total number of h-refinements: " << numH
-           << "\nTotal number of p-refinements: " << numP
-           << "\nMaximum order " << maxP << "\n";
-   }
+   const int size = fespace.GetTrueVSize();
+   cout << "Number of finite element unknowns: " << size << endl;
 
-   ParGridFunction x(&fespace);
+   const int maxP = fespace.GetMaxElementOrder();
+   cout << "Total number of h-refinements: " << numH
+        << "\nTotal number of p-refinements: " << numP
+        << "\nMaximum order " << maxP << "\n";
+
+   GridFunction x(&fespace);
    Vector X;
 
    if (projectSolution)
@@ -202,90 +168,88 @@ int main(int argc, char *argv[])
 
       X.SetSize(fespace.GetTrueVSize());
 
-      fespace.GetRestrictionMatrix()->Mult(x, X);
+      fespace.GetHpRestrictionMatrix()->Mult(x, X);
       fespace.GetProlongationMatrix()->Mult(X, x);
 
       // Compute and print the L^2 norm of the error.
       const real_t error = x.ComputeL2Error(vec_coef);
-      if (myid == 0)
-      {
-         cout << "\n|| E_h - E ||_{L^2} = " << error << '\n' << endl;
-      }
+      cout << "\n|| E_h - E ||_{L^2} = " << error << '\n' << endl;
    }
    else
    {
-      // 8. Determine the list of true (i.e. parallel conforming) essential
-      //    boundary dofs. In this example, the boundary conditions are defined
-      //    by marking all the boundary attributes from the mesh as essential
-      //    (Dirichlet) and converting them to a list of true dofs.
+      // 6. Determine the list of essential boundary dofs. In this example, the
+      //    boundary conditions are defined by marking all the boundary attributes
+      //    from the mesh as essential (Dirichlet) and converting them to a list of
+      //    true dofs.
       Array<int> ess_tdof_list;
-      if (pmesh.bdr_attributes.Size())
+      if (mesh.bdr_attributes.Size())
       {
-         Array<int> ess_bdr(pmesh.bdr_attributes.Max());
+         Array<int> ess_bdr(mesh.bdr_attributes.Max());
          ess_bdr = 1;
          fespace.GetEssentialTrueDofs(ess_bdr, ess_tdof_list);
       }
 
-      // 9. Set up the parallel linear form b(.) which corresponds to the
-      //    right-hand side of the FEM linear system, which in this case is
-      //    (1,phi_i) where phi_i are the basis functions in fespace.
-      ParLinearForm b(&fespace);
+      // 7. Set up the linear form b(.) which corresponds to the right-hand side of
+      //    the FEM linear system, which in this case is (1,phi_i) where phi_i are
+      //    the basis functions in fespace.
+      LinearForm b(&fespace);
       ConstantCoefficient one(1.0);
       b.AddDomainIntegrator(new DomainLFIntegrator(one));
       b.Assemble();
 
-      // 10. Define the solution vector x as a parallel finite element grid
-      //     function corresponding to fespace. Initialize x with initial guess of
-      //     zero, which satisfies the boundary conditions.
+      // 8. Define the solution vector x as a finite element grid function
+      //    corresponding to fespace. Initialize x with initial guess of zero,
+      //    which satisfies the boundary conditions.
       x = 0.0;
 
-      // 11. Set up the parallel bilinear form a(.,.) on the finite element space
-      //     corresponding to the Laplacian operator -Delta, by adding the
-      //     diffusion domain integrator.
-      ParBilinearForm a(&fespace);
+      // 9. Set up the bilinear form a(.,.) on the finite element space
+      //    corresponding to the Laplacian operator -Delta, by adding the diffusion
+      //    domain integrator.
+      BilinearForm a(&fespace);
       a.AddDomainIntegrator(new DiffusionIntegrator(one));
 
-      // 12. Assemble the parallel bilinear form and the corresponding linear
-      //     system, applying any necessary transformations such as: parallel
-      //     assembly, eliminating boundary conditions, applying conforming
-      //     constraints for non-conforming AMR, static condensation, etc.
+      // 10. Assemble the bilinear form and the corresponding linear system,
+      //     applying any necessary transformations such as: assembly, eliminating
+      //     boundary conditions, applying conforming constraints for non-conforming
+      //     AMR, static condensation, etc.
       a.Assemble();
 
       OperatorPtr A;
       Vector B;
       a.FormLinearSystem(ess_tdof_list, x, b, A, X, B);
 
-      // 13. Solve the linear system A X = B.
-      //     * With full assembly, use the BoomerAMG preconditioner from hypre.
-      //     * With partial assembly, use Jacobi smoothing, for now.
-      Solver *prec = new HypreBoomerAMG;
-      CGSolver cg(MPI_COMM_WORLD);
-      cg.SetRelTol(1e-12);
-      cg.SetMaxIter(2000);
-      cg.SetPrintLevel(1);
-      if (prec) { cg.SetPreconditioner(*prec); }
-      cg.SetOperator(*A);
-      cg.Mult(B, X);
-      delete prec;
+      // 11. Solve the linear system A X = B.
+      {
+#ifndef MFEM_USE_SUITESPARSE
+         // Use a simple symmetric Gauss-Seidel preconditioner with PCG.
+         GSSmoother M((SparseMatrix&)(*A));
+         PCG(*A, M, B, X, 1, 200, 1e-12, 0.0);
+#else
+         // If MFEM was compiled with SuiteSparse, use UMFPACK to solve the system.
+         UMFPackSolver umf_solver;
+         umf_solver.Control[UMFPACK_ORDERING] = UMFPACK_ORDERING_METIS;
+         umf_solver.SetOperator(*A);
+         umf_solver.Mult(B, X);
+#endif
+      }
 
-      // 14. Recover the parallel grid function corresponding to X. This is the
-      //     local finite element solution on each processor.
+      // 12. Recover the grid function corresponding to X.
       a.RecoverFEMSolution(X, b, x);
    }
 
    if (fespaceDim == 1)
    {
       const real_t h1error = CheckH1Continuity(x);
-      cout << myid << ": H1 continuity error " << h1error << endl;
+      cout << "H1 continuity error " << h1error << endl;
       MFEM_VERIFY(h1error < 1.0e-12, "H1 continuity is not satisfied");
    }
 
    L2_FECollection fecL2(0, dim);
-   ParFiniteElementSpace l2fespace(&pmesh, &fecL2);
-   ParGridFunction xo(&l2fespace);
+   FiniteElementSpace l2fespace(&mesh, &fecL2);
+   GridFunction xo(&l2fespace);
    xo = 0.0;
 
-   for (int e=0; e<pmesh.GetNE(); ++e)
+   for (int e=0; e<mesh.GetNE(); ++e)
    {
       const int p_elem = fespace.GetElementOrder(e);
       Array<int> dofs;
@@ -294,41 +258,30 @@ int main(int argc, char *argv[])
       xo[dofs[0]] = p_elem;
    }
 
-   // 15. Save the refined mesh and the solution in parallel. This output can
-   //     be viewed later using GLVis: "glvis -np <np> -m mesh -g sol".
+   // 13. Save the refined mesh and the solution. This output can be viewed later
+   //     using GLVis: "glvis -m refined.mesh -g sol.gf".
    std::unique_ptr<GridFunction> vis_x = x.ProlongateToMaxOrder();
-   {
-      ostringstream mesh_name, sol_name, order_name;
-      mesh_name << "mesh." << setfill('0') << setw(6) << myid;
-      sol_name << "sol." << setfill('0') << setw(6) << myid;
-      order_name << "order." << setfill('0') << setw(6) << myid;
+   ofstream mesh_ofs("refined.mesh");
+   mesh_ofs.precision(8);
+   mesh.Print(mesh_ofs);
+   ofstream sol_ofs("sol.gf");
+   sol_ofs.precision(8);
+   vis_x->Save(sol_ofs);
+   ofstream order_ofs("order.gf");
+   order_ofs.precision(8);
+   xo.Save(order_ofs);
 
-      ofstream mesh_ofs(mesh_name.str().c_str());
-      mesh_ofs.precision(8);
-      pmesh.ParPrint(mesh_ofs);
-
-      ofstream sol_ofs(sol_name.str().c_str());
-      sol_ofs.precision(8);
-
-      vis_x->Save(sol_ofs);
-
-      ofstream order_ofs(order_name.str().c_str());
-      order_ofs.precision(8);
-      xo.Save(order_ofs);
-   }
-
-   // 16. Send the solution by socket to a GLVis server.
+   // 14. Send the solution by socket to a GLVis server.
    if (visualization)
    {
       char vishost[] = "localhost";
       int  visport   = 19916;
       socketstream sol_sock(vishost, visport);
-      sol_sock << "parallel " << num_procs << " " << myid << "\n";
       sol_sock.precision(8);
-      sol_sock << "solution\n" << pmesh << *vis_x << flush;
+      sol_sock << "solution\n" << mesh << *vis_x << flush;
    }
 
-   // 17. Free the used memory.
+   // 15. Free the used memory.
    if (delete_fec)
    {
       delete fec;
@@ -337,19 +290,15 @@ int main(int argc, char *argv[])
    return 0;
 }
 
-real_t CheckH1Continuity(ParGridFunction & x)
+real_t CheckH1Continuity(GridFunction & x)
 {
-   x.ExchangeFaceNbrData();
-
-   const ParFiniteElementSpace *fes = x.ParFESpace();
-   ParMesh *mesh = fes->GetParMesh();
+   const FiniteElementSpace *fes = x.FESpace();
+   Mesh *mesh = fes->GetMesh();
 
    const int dim = mesh->Dimension();
 
    // Following the example of KellyErrorEstimator::ComputeEstimates(),
-   // we loop over interior faces and then shared faces.
-
-   // Compute error contribution from local interior faces
+   // we loop over interior faces and compute their error contributions.
    real_t errorMax = 0.0;
    for (int f = 0; f < mesh->GetNumFaces(); f++)
    {
@@ -389,34 +338,6 @@ real_t CheckH1Continuity(ParGridFunction & x)
                errorMax = std::max(errorMax, err_i);
             }
          }
-      }
-   }
-
-   // Compute error contribution from shared interior faces
-   for (int sf = 0; sf < mesh->GetNSharedFaces(); sf++)
-   {
-      const int f = mesh->GetSharedFace(sf);
-      const bool trueInterior = mesh->FaceIsTrueInterior(f);
-      if (!trueInterior) { continue; }
-
-      auto FT = mesh->GetSharedFaceTransformations(sf, true);
-      const int faceOrder = dim == 3 ? fes->GetFaceOrder(f) : fes->GetEdgeOrder(f);
-      const auto &int_rule = IntRules.Get(FT->FaceGeom, 2 * faceOrder);
-      const auto nip = int_rule.GetNPoints();
-
-      for (int i = 0; i < nip; i++)
-      {
-         const auto &fip = int_rule.IntPoint(i);
-         IntegrationPoint ip;
-
-         FT->Loc1.Transform(fip, ip);
-         const real_t v1 = x.GetValue(FT->Elem1No, ip);
-
-         FT->Loc2.Transform(fip, ip);
-         const real_t v2 = x.GetValue(FT->Elem2No, ip);
-
-         const real_t err_i = std::abs(v1 - v2);
-         errorMax = std::max(errorMax, err_i);
       }
    }
